Make sync execute wrap async
